refactor(carRacing) : package를 model과 view로 재구성
