+++ src/main/java/com/java/jaeheon/step2/RacingGame.java
@@ -1,5 +1,10 @@
 package com.java.jaeheon.step2;
 
+import com.java.jaeheon.step2.model.CarManagement;
+import com.java.jaeheon.step2.model.CarRacing;
+import com.java.jaeheon.step2.view.InputView;
+import com.java.jaeheon.step2.view.ResultView;
+
 public class RacingGame {
 
     public static void main(String[] args) {
