+++ src/test/java/com/java/jaeheon/step2/model/CarManagementTest.java
@@ -1,9 +1,11 @@
-package com.java.jaeheon.step2;
+package com.java.jaeheon.step2.model;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import com.java.jaeheon.step2.model.CarManagement;
 
 import static org.assertj.core.api.Assertions.assertThat;
