+++ src/test/java/com/java/jaeheon/step2/model/CarTest.java
@@ -1,8 +1,10 @@
-package com.java.jaeheon.step2;
+package com.java.jaeheon.step2.model;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import com.java.jaeheon.step2.model.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
